+++ site/dat/docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.4.3 (next)
  - allow passing config as stdin to CLI like `./generate-config.sh | bzt`
+ - allow having `report-name=ask` for BlazeMeter reporter
 
 ## 0.4.2
  - bump up Gatling version to 2.1.7